Make ul.operator-list its own memoized component
Opening the menus is slow because it causes a rerender of the operator
list, even though nothing has changed. Memoize it so that it doesn't
have to render in this case.
